Rename `sync` mod to `lock` mod
