+++ examples/rwlock_write/src/lib.rs
@@ -1,7 +1,7 @@
 #![no_std]
 
 use core::num::ParseIntError;
-use gstd::sync::rwlock::RwLock;
+use gstd::lock::rwlock::RwLock;
 use gstd::{exec, msg, prelude::*, ActorId};
 
 static mut PING_DEST: ActorId = ActorId::new([0u8; 32]);
